nvmf_tgt: wait for inflight IOs in disconnect-host flow
This patch adds a mechanism to wait for all outstanding IOs during the
disconnect-host-from-subsystem flow. As a result, when the operation
completes successfully, the caller is guaranteed that there are no
inflight IOs from the disconnected host in the subsystem. This prevents
potential data corruption and avoids undesirable behavior from the
host/cluster point of view.

Because the time required for all inflight IOs to finish can depend on
the subsystem configuration, the number of outstanding operations, and
the host/cluster state, the new implementation allows specifying a
custom timeout for each disconnect-host operation. If the function is
called with a timeout value of 0, the default timeout
NVMF_CTRLR_RESET_SHN_TIMEOUT_IN_MS is used.

Change-Id: Ief4accda8fb5718b4a7e6ddff0e5a8d130c4b462
Signed-off-by: Marcin Dziegielewski <marcin.dziegielewski@dell.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/26389
Reviewed-by: Jim Harris <jim.harris@nvidia.com>
Reviewed-by: Jacek Kalwas <jacek.kalwas@nutanix.com>
Community-CI: Mellanox Build Bot
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>