test/common: "Workaround" Bison's line break offset
Bash uses Bison for its grammar and there's a small, existing
quirk which may make traces a bit confusing in context of how
backslash-line continuations are parsed. In case error is trapped
by the following cmd:

  false \
    foo \
    bar

trace info will have its line pointer set to

  foo \

instead of

 false \

Essentially, when backslashes are used the continuation eats the newline
but the lexer still increments the internal line counter and sets token
location to the current line - this is now the value tracing will hold on
to (the increment happens as technically, at the very end, this is still
a valid line).

In practice, the following example:

  a() { b foo bar; }
  b() { c bar foo; }
  c() { d "Cake is a lie"; }
  d() { false \
  "oh, is it really here?" \
  "or maybe here?"
  }
  a "Let's propagate errors" "in automated ways"

Yields:

  frame_id  func  func_args         source           line
  [0]       d     ("Cake is a lie") ./bt.bash:9  "oh, is it really here?" \
  ...

With the proposed workaround (assuming code is properly formatted as
per our check_format|shfmt) it should become:

  frame_id  func  func_args         source           line
  [0]       d     ("Cake is a lie") ./bt.bash:8   d() { false \

Change-Id: I1aca294f98bb62e3e07bc2c41f1aa3a1976f9337
Signed-off-by: Michal Berger <michal.berger@nutanix.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/27109
Reviewed-by: Konrad Sztyber <ksztyber@nvidia.com>
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Jim Harris <jim.harris@nvidia.com>