test/common: Add '+' at the beginning of the PS4
PS4 is used together with tracing and its first character is
supposed to be replicated multiple times to indicate proper levels
of indirection (BASH(1)).

Since in our case this was a space said indirections where somewhat
"invisible", in fact, unless user is fully aware of PS4's usage, it
may come across as simply broken formatting with "random" number
of prefixed spaces.

'+' is a default char used for indicating these levels properly
so let's use it for clarity.

Change-Id: If609e21df92d621d5aeea4a4f2e6cc0fda351947
Signed-off-by: Michal Berger <michal.berger@nutanix.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/27110
Reviewed-by: Krzysztof Goreczny <krzysztof.goreczny@dell.com>
Reviewed-by: Jim Harris <jim.harris@nvidia.com>
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Karol Latecki <karol.latecki@nutanix.com>
Reviewed-by: Konrad Sztyber <ksztyber@nvidia.com>