test/common: Add coloring support to backtraces
Introduce some helpers for coloring the output. This is done to,
hopefully, enhance visibility of failed components (under terminals,
log viewers which are capable of decoding color codes).

For now, only basic palette is added, can be extended in the future.
By default, coloring is disabled, AUTOTEST_COLORS_ENABLED=true can
be used to enable it.

Use it to make the failing frame in the backtrace visible as red in
case coloring is enabled.

Change-Id: I7cf3fc03f6204390fb354fde77982c7ed83b5139
Signed-off-by: Michal Berger <michal.berger@nutanix.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/27135
Reviewed-by: Konrad Sztyber <ksztyber@nvidia.com>
Reviewed-by: Jim Harris <jim.harris@nvidia.com>
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>