test/common: Add wrapper around date to fetch unix epoch
Since use of date may be not compatible across different OSs, use
Bash's env var as a default instead. Put it into function and replace
all direct calls to date.

As we are at it, use it to calculate time it took to dump the
backtrace summary (i.e. time it took to perform cleanup after
failure).

Change-Id: I250898f0014fa141daf7f1252dc09352bcdcdfc3
Signed-off-by: Michal Berger <michal.berger@nutanix.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/27211
Reviewed-by: Jim Harris <jim.harris@nvidia.com>
Reviewed-by: Konrad Sztyber <ksztyber@nvidia.com>
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Tomasz Zawadzki <tomasz@tzawadzki.com>