pkgdep/rhel: use baseurl instead mirrorlist for elrepo
ElRepo mirror is notoriously unstable and tends to time
out, making pkgdep.sh (and the CI in some cases) fail.

The reason is not 100% clear.
Some traces "in the internet" suggest denial of service,
e.g.: https://github.com/elrepo/packages/issues/242

Other suggest that some public network prefixes attempting
to access the mirror might be (at least temporarily) banned
due to excessive number of connections:
https://lowendtalk.com/discussion/182220/failed-reaching-http-mirrors-elrepo-org-mirrors-elrepo-el8-from-racknerd-vps

As suggested in one of the links - switch to using baseurl
instead of using mirrorlist. The switch will be done only if
/etc/yum.repos.d/elrepo.conf is added for the first time.
If it's already present then it's likely we're on someone's
dev system, so better not touch the configuration.

Change-Id: I2f8d5c4362f648d7d6d0fd48446e789e3d10b957
Signed-off-by: Karol Latecki <karol.latecki@nutanix.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/27354
Reviewed-by: Tomasz Zawadzki <tomasz@tzawadzki.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Jim Harris <jim.harris@nvidia.com>
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>